+++ src/test/java/nextstep/subway/acceptance/StationAcceptanceTest.java
@@ -12,7 +12,7 @@ import org.springframework.http.HttpStatus;
 
 import java.util.List;
 
-import static nextstep.subway.acceptance.StationUtils.*;
+import static nextstep.subway.acceptance.utils.StationUtils.*;
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.junit.jupiter.api.Assertions.assertAll;
