+++ src/test/java/nextstep/subway/acceptance/utils/StationUtils.java
@@ -1,4 +1,4 @@
-package nextstep.subway.acceptance;
+package nextstep.subway.acceptance.utils;
 
 import io.restassured.RestAssured;
 import io.restassured.response.ExtractableResponse;
